doc: Replace grass.script by grass.tools in intros (#6111)
Replaces grass.script run_command calls by grass.tools in the main documentation (Markdown intro and overview) and adds it to Python doc (Sphinx rst).

Change the structure (headings) to focus on the use case, not the style of API. Splits and renames the part about grass.pygrass.

Replaces run_command by grass.tools in Jupyter examples, but with limited testing because that tutorial needs overall updates and polish.