tests: Enable tests for gunittest's test_assertions_vect.py (#6124)
I was preparing some changes in gunittest code, and noticed that a whole set of asserts were not tested, and there was a typo in one of them for more than a decade. Unfortunately, the tests here weren't checking for this particular assert.
The file was excluded because of one error that is unrelated to the problem I found.
By making it an expected failure, the 10 other tests in the file can be executed.


* grass.gunittest: Mark test_assertVectorEqualsAscii_by_import as expectedFailure to enable running the remaining of the file

* tests: Enable tests for gunittest's test_assertions_vect.py