v.decimate: Add tests (#6152)
Adds tests covering different parameters of the tool, although testing only the basics for the outputs.

Uses the new Tools API which reduced handling of stdout and stderr outputs to minimum and does not create need for wrapper functions or temporary variables. Uses SimpleNamespace to pass things from the fixture to the tests.