grass.pygrass: Remove Vector.write() 2nd positional argument deprecated in 7.2 (#6231)
Removes the compatibility code in Vector.write() which accepts the 7.0 API positional argument in 7.2 and above. It has already been a decade since that workaround to keep the calling syntax of GRASS 7.0 and before valid and we have a new major version already established.

Fixes v.rast.stats tests which is using the old signature.