v.db.connect: Add JSON support (#6077)
This PR adds JSON support to the v.db.connect tool. The JSON output has a list as the first level item because the tool gives connections with -p, and we don't anticipate there is anything to output for the whole vector map: one or more connections are associated with a map while per-map information is provided with v.info. CSV simply follows the JSON while providing a legacy variant.

The column info with -c is printed for a specific layer, so a CSV with columns for names and types is used which roughly corresponds to the original, now legacy output which is still provided.

---------

Signed-off-by: Nishant Bansal <nishant.bansal.282003@gmail.com>