v.colors: Fix colors returning to default in another mapsets (#6240)
This fixes the color table lookup code so that if mapset is not the current mapset and the local vcolr2 table is not defined, GRASS will fallback to using the color table in the original mapset (as opposed to falling back to default/no color table). Both the behavior and the code follows the raster library.