doc: Use CRS instead of SRS
We use both CRS and SRS in the documentation. There is more CRS than SRS, so this replaces the few occurences of SRS with CRS. In case of WMS, also mention SRS because the SRS/srs parameter is explicitly mentioned. In general, we can't avoid SRS because it is used a lot in the code, in GDAL, and at this point in tool parameters as well. This does not touch the cases where 'projection' is used to refer to CRS.