v.what: Expand doc to describe outputs, JSON
This expands documentation in light of the new JSON addition (not merged at the time of writing), but also generally describing how the tool behaves. Additionally, this improves how -a and -m flags are described and presented. columns now requires -a because it doesn't do anything otherwise. r.what is linked from doc.