g.list: Print empty list for JSON (#6302)
If nothing is found, still print an empty JSON list with format=json (but still nothing with other formats). Additionally, sort only with list with two or more items. While calling qsort on zero items should work, there is no reason to call it on zero items or one item.