v.what: Expand doc to describe outputs, JSON, flags (#6266)
This expands documentation in light of the new JSON addition, but also generally describes how the tool behaves. Additionally, this improves how -a and -m flags are described and presented. The columns parameters now requires -a because it doesn't do anything otherwise. The r.what tool is now linked from doc.