lib/raster: Add bad line to reclass header error
When a line has an unexpected format when reading reclass map header file, a warning is issued (before the fatal error), but the warning does not give any details. This adds the problematic line to the warning by passing it to the caller through an optional pointer to a string parameter.

A pytest test which tests the normal behavior and also specifically the error is added. An additional grass.gunittest is included which more closely copies the now problematic r.random.surface test with mask. The tests may need adjusting based on how CI look like on macOS.