lib/raster: Add bad line to reclass header error (#6339)
When a line has an unexpected format when reading reclass map header file, a warning is issued (before the fatal error), but the warning does not give any details. This adds the problematic line to the warning by passing it to the caller through an optional pointer to a string parameter. The reported error message now reports details about the parsing state.

A pytest test which tests the normal behavior and also pytest tests specifically for the error states are added.

Also, make the existing r.random.surface with mask a separate test file (as an alternative to a separate test which is trying to be similar and also to avoid wondering if there is any mask interaction with the other test methods).