lib/raster: Add larger buffers, error messages to reclass header reader (#6341)
This increases the size of buffer the reclass header reader code (reading header of the header file) uses for reading raster and mapset names. It also adds an additional check for reading the whole line and errors if not read. A message can now be optionally reported for each of the already handled error states.

The maximum length of files is now checked by a test which fails with the original buffer size combined with both the original code and the new error handling code, but the individual error states are not tested.

Using 50 as the name length on Windows to avoid 'MAPSET m... not found at' error from r.mapcalc call which occurs with length 255.