r.pack: Replace traceback by error message for output issues
Add check for issues with the output file parent directory, and report error messages. This avoids a traceback originating at the very end of the tool execution when calling shutil.move.

This adds tests for the newly handled error states as well as for other issues with the inputs.

It adds the new pytest fixtures to scripts conftest.py file given there is at least some potential for reusability for similar simple tests.