docker: Change Debian base image to 12.5-slim (#6373)
The current image is still pinned to an image generated on June 2025. We are not getting new Renovate PRs for it, as the suffix is not recognized as a version. Since we are not pinning the image disgest, we can use a numeric prefix that will be recognized by Renovate. I chose 12.5-slim, instead of bookworm-slim or 12-slim, as I think it is easier to understand what is changed when receiving a PR, that would only contain a digest update otherwise.