grass.script: Remove build vars from setup.py
Remove build vars from setup and use the existing dependency on grass.app.runtime to get the information. Additionally, move some of the parametrization to grass.app.runtime because now, both the functions executing the setup steps and the variable are there, so there is no reason to supply these from the outside.

Config dir does not require version, but I kept a general function there which is used in g.extension and in GUI because these use the runtime-obtained version as opposed to the one from grass.app.runtime (difference between running system and setup of the env).

I removed the build variables from the documentation which makes it a little simpler to edit, although less explicit assuming that the version usage there was actually correct. The example needs more unrelated edits, so I'm keeping it as is.