grass.tools: Use new ToolError, not CalledModuleError (#6366)
The new ToolError tries to be even more aligned with CalledProcessError (by focusing on cmd and not tool and code like CalledModuleError), while providing messages more tailored towards Tools object and use cases. It also tries to polish further short one-line messages and distinguish between empty stderr and not captures stderr. API and documentation of it is more narrow and focused on the Tools use case specifically and includes type annotations.

Document exception also in each tool page in the Parameters section followed by Returns: and Raises: blocks. The new Raises: block again mimics docs like NumPy (like the Returns: block), but with some liberty (it is in fact under section Parameters and it uses bold, emphasis and indent differently because of consistency with how we want the parameters to be represented and technical limitations related to indents which we do manually). The "tool" terminology now creates a consistent result in the doc (CalledModuleError was not on the tool pages because it would not be consistent).