g.region: Output CRS info in JSON
The original g.region output contains basic info about the current CRS (aka projection). While this is only basic, providing only distinction between XY, LL, and other (and UTM+zone to certain extent), it is used in the code base and expected to be correct, for example g.proj is using it before checking anything else. Although it is separate from the stored PROJ files, it serves as the primary way of checking what kind of CRS is one dealing with. While this is unrelated to computational region, and in fact, it cannot be modified by g.region (it is ready-only for g.region), you could say it encourages a correct interpretation of the computational region information provided by g.region.

This adds pytest-based tests covering XY, LL, UTM N, UTM S, and other CRS cases.

The key names, the structure, and the values are simply taken from the current output with format='shell', i.e., keys are at the top-level, called 'projection' and 'zone', and values are integers. This means that there is no obvious way how to interpret that information, but the information can be directly used to compare with another computational region even when accessed in a different way (like directly looking into the DEFAULT_WIND file).