docker: Remove scikit-learn from Alpine test (#6394)
The Docker test with Alpine is using scikit-learn, presuably to support the test of installing r.learn.ml2 (which in theory should install without it). This changes the test to use r.diversity instead of r.learn.ml2 which does not come with any dependency, reducing the setup and teardown phases, while still testing installation of a Python tool from grass-addons.