r.pack: Store all CRS info in the file (#6408)
The r.pack missed couple of the new PROJ-underscore files which are describing or defining the CRS (aka projection) at the project level. Specifically, projects created with EPSG have PROJ_SRID file (not PROJ_EPSG) which was previously omitted, and thus missing in the exported raster pack file. Additionally, the implementation missed another project-level information which is stored in and retrieved from the default computational region, namely projection and zone keys. The new code reads them with g.region and stores them in a JSON file inside the pack file.

The new tests check for the expected files and check the format and content of the newly added JSON file.