lib/external/parson: hide parson symbols on linkage
Various open source JSON libraries in use, share in part identical API.
In general this is because they do not use library specific prefixes.
This can cause linkage conflicts, as has been the case with GDAL linked
a call to the GRASS parson library instead of its own embedded JSON-C,
causing segfault.

This fix includes a minor patch to the vendored parson code, which
would have been avoided if there was a simple way to pass a CFLAG
"-fvisibility=hidden" to a single file in GRASS' build system.