lib/external/parson: hide parson symbols on linkage (#6432)
Various open source JSON libraries in use, share in part identical API.
In general this is because they do not use library specific prefixes.
This can cause linkage conflicts, as has been the case with GDAL linked
a call to the GRASS parson library instead of its own embedded JSON-C,
causing segfault.

This fix includes a minor patch to the vendored parson code, which
would have been avoided if there was a simple way to pass a CFLAG
"-fvisibility=hidden" to a single file in GRASS' build system.

* Add test for r.out.gdal based on the code provided in the original
  report in #6398 (export of COG segfaults). The tests fails with -11
  without the changes and passes with the changes.

---------

Co-authored-by: Vaclav Petras <wenzeslaus@gmail.com>